python/ml-requirements.txt
# Core PyTorch + Lightning
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
pytorch-lightning==2.5.6
torchmetrics==1.8.2
triton==3.5.0

# ONNX + conversion tools
onnx==1.16.1
onnxruntime==1.17.3
onnx-tf==1.10.0
onnx-simplifier==0.4.36
sng4onnx==1.0.4
onnx-graphsurgeon==0.5.2
onnxscript==0.5.6

# Ray (for hyperparameter tuning / distributed training)
ray[tune]


# Misc. utilities
numpy
scipy
pandas
pydantic
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
typing-extensions==4.15.0